Add image save functionality

+++ src/main/java/com/farmtrade/dto/product/CreateProductDto.java
@@ -5,12 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class CreateProductDto {
-    private Float quantity;
-    private Float size;
+    private BigDecimal quantity;
+    private BigDecimal size;
     private Long productName;
 }
